+++ packages/components/ui/src/index.js
@@ -23,4 +23,5 @@ export {default as NoScript} from './NoScript.svelte';
 export {default as ResponsiveButton} from './ResponsiveButton.svelte';
 export {default as ResponsiveFlex} from './ResponsiveFlex.svelte';
 export {default as ScrollbarStyler} from './ScrollbarStyler.svelte';
+export {default as Scroller} from './Scroller.svelte';
 export {default as Switch} from './Switch.svelte';